+++ src/main/java/com/liquibaseproject/service/UsersService.java
@@ -1,6 +1,7 @@
 package com.liquibaseproject.service;
 
 import com.liquibaseproject.entity.Users;
+import com.liquibaseproject.exception.ResultsNotFoundException;
 import com.liquibaseproject.repository.UsersRepository;
 import io.micrometer.common.util.StringUtils;
 import org.springframework.stereotype.Service;
@@ -15,12 +16,18 @@ public class UsersService {
 
     private final UsersRepository usersRepository;
 
+    private static final String USER_NOT_FOUND_MESSAGE = "User with ID %s does not exist.";
+
     public UsersService(UsersRepository usersRepository) {
         this.usersRepository = usersRepository;
     }
 
     public Optional<Users> getUserById(UUID id) {
-        return usersRepository.findById(id);
+        if (usersRepository.existsById(id)) {
+            return usersRepository.findById(id);
+        } else {
+            throw new ResultsNotFoundException(String.format(USER_NOT_FOUND_MESSAGE, id));
+        }
     }
 
     public List<Users> findByUsernameIgnoreCase(String userName) {
@@ -47,10 +54,12 @@ public class UsersService {
 
     public void updateUser(UUID id, Users userEntity) {
         Users user = usersRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("User not found"));
+                .orElseThrow(() -> new ResultsNotFoundException(String.format(USER_NOT_FOUND_MESSAGE, id)));
+
         if (StringUtils.isNotEmpty(userEntity.getUsername())) {
             user.setUsername(userEntity.getUsername());
         }
+
         if (StringUtils.isNotEmpty(userEntity.getEmail())) {
             user.setEmail(userEntity.getEmail());
         }
@@ -59,6 +68,10 @@ public class UsersService {
     }
 
     public void deleteUser(UUID id) {
-        usersRepository.deleteById(id);
+        if (usersRepository.existsById(id)) {
+            usersRepository.deleteById(id);
+        } else {
+            throw new ResultsNotFoundException(String.format(USER_NOT_FOUND_MESSAGE, id));
+        }
     }
 }
